+++ packages/next-sdk/index.ts
@@ -57,17 +57,4 @@ export {
 
 export * from '@mcp-b/webmcp-polyfill'
 
-import { initializeWebMCPPolyfill } from '@mcp-b/webmcp-polyfill'
-import { setupModelContextBridge } from './page-tools/bridge'
-import { isBrowser } from './utils/env'
-
-export const initializeBuiltinWebMCP = () => {
-  if (isBrowser()) {
-    try {
-      initializeWebMCPPolyfill()
-      setupModelContextBridge()
-    } catch (err) {
-      console.warn('[next-sdk] 自动注入 modelContext polyfill 和桥接同步失败:', err)
-    }
-  }
-}
+export { initializeBuiltinWebMCP } from './page-tools/initialize-builtin-WebMCP'
